## New Files Created
- `PyPacketValidator.h` - Unified packet validation framework header file
- `PyPacketValidator.cpp` - Packet validator implementation
## Enhanced Validation
1.1.
Tuple Validation - Strictly checks tuple size and element non-nullability
2.2.
Type Validation - Ensures PyRep object types match expectations
3.3.
String Validation - Checks string length and safety (prevents null byte injection)
4.4.
Integer Range Validation - Verifies that values ​​are within a reasonable range
5.5.
Packet Structure Integer Integer - Ensures packets are valid stream types
6.6.
Debug Output - Unified error logging and packet content dumping
## Modified Files
- `PyPacket.cpp` - Integrates new validator, replacing existing validation logic
- `CMakeLists.txt` - Adds new source files to the build system
## Improved Decoding Methods
- `PyPacket::Decode`
- `PyAddress::Decode`
- `PyCallStream::Decode`
These improvements significantly improve the accuracy, error handling, and debugging of EVE network packet parsing, reducing the risk of server crashes caused by malformed packets. The project compiles successfully, and all validation features have been integrated into the existing code.

+++ src/eve-server/inventory/ItemFactory.h
@@ -57,7 +57,7 @@ public:
 
     void Close();
     int Initialize();
-    uint32 Count()                                      { return m_items.size(); }
+    uint32 Count()                                      { return static_cast<uint32>(m_items.size()); }
 
     void SaveItems();
     void RemoveItem(uint32 itemID);
