+++ src/eve-server/system/SystemBubble.h
@@ -77,8 +77,8 @@ public:
 
     /* various count queries */
     uint32 CountNPCs();
-    uint32 CountPlayers()                               { return m_players.size(); }
-    uint32 CountDynamics()                              { return m_dynamicEntities.size(); }
+    uint32 CountPlayers()                               { return static_cast<uint32>(m_players.size()); }
+    uint32 CountDynamics()                              { return static_cast<uint32>(m_dynamicEntities.size()); }
 
     /* used for bubble management */
     bool IsEmpty() const                                { return (m_entities.empty() ? m_dynamicEntities.empty() : false); }
